# e-commerce_backend/openapi-spec.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/admin": {
            "get": {
                "operationId": "AdminController_getAdminInfo",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Admin"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/admin/getadmin": {
            "get": {
                "operationId": "AdminController_getAdminNameandId",
                "parameters": [
                    {
                        "name": "name",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Admin"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/admin/addadmin": {
            "post": {
                "operationId": "AdminController_addAdmin",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Admin"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/admin/delete/{id}": {
            "delete": {
                "operationId": "AdminController_deleteAdmin",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Admin"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/admin/getadmin/{id}": {
            "put": {
                "operationId": "AdminController_updateAdmin",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Admin"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            },
            "patch": {
                "operationId": "AdminController_patchAdmin",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Admin"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/admin/sellers/pending": {
            "get": {
                "operationId": "AdminController_getPendingSellers",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Admin"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/admin/sellers/verified": {
            "get": {
                "operationId": "AdminController_getVerifiedSellers",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Admin"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/admin/sellers/{id}/verify": {
            "post": {
                "operationId": "AdminController_verifySeller",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifySellerDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Admin"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/admin/sellers/{id}/reject": {
            "post": {
                "operationId": "AdminController_rejectSeller",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RejectSellerDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Admin"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/admin/orders": {
            "get": {
                "operationId": "AdminController_getOrders",
                "parameters": [
                    {
                        "name": "page",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "status",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Admin"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/admin/orders/{id}/status": {
            "patch": {
                "operationId": "AdminController_updateOrderStatus",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateOrderStatusDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Admin"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/admin/emails/send": {
            "post": {
                "operationId": "AdminController_sendEmail",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendEmailDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Admin"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/admin/emails/send-bulk": {
            "post": {
                "operationId": "AdminController_sendBulkEmail",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendBulkEmailDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Admin"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/admin/emails/history": {
            "get": {
                "operationId": "AdminController_getEmailHistory",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Admin"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/admin/dashboard/trends": {
            "get": {
                "operationId": "AdminController_getDashboardTrends",
                "summary": "Get dashboard trend data for charts",
                "description": "Returns time-series data for users, sellers, and products over a specified period",
                "parameters": [
                    {
                        "name": "period",
                        "required": false,
                        "in": "query",
                        "description": "Time period for trends (7days, 30days, 3months, 1year)",
                        "schema": {
                            "enum": [
                                "7days",
                                "30days",
                                "3months",
                                "1year"
                            ],
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trends data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DashboardTrendsResponseDto"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing token"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "tags": [
                    "Admin"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
...
        "/api/v1/orders": {
            "post": {
                "operationId": "OrderController_createOrder",
                "summary": "Create a new order",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully"
                    },
                    "400": {
                        "description": "Invalid order data"
                    },
                    "403": {
                        "description": "Authentication required"
                    }
                },
                "tags": [
                    "Orders"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            },
            "get": {
                "operationId": "OrderController_getAllOrders",
                "parameters": [
                    {
                        "name": "page",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Orders"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
...
        "/api/v1/orders/{id}/status": {
            "patch": {
                "operationId": "OrderController_updateOrderStatus",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateOrderStatusDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Orders"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "E-Commerce Platform API",
        "description": "Complete API documentation for the E-Commerce platform with authentication, products, orders, notifications, and seller management.",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [
        {
            "url": "http://localhost:4002",
            "description": "Development Server"
        },
        {
            "url": "http://localhost:4002/api/v1",
            "description": "API v1"
        }
    ],
    "components": {
        "securitySchemes": {
            "JWT-auth": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http",
                "description": "Enter JWT token"
            },
            "cookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "access_token",
                "description": "JWT token stored in HTTP-only cookie"
            }
        },
        "schemas": {
            "VerifySellerDto": {
                "type": "object",
                "properties": {}
            },
            "RejectSellerDto": {
                "type": "object",
                "properties": {}
            },
            "UpdateOrderStatusDto": {
                "type": "object",
                "properties": {}
            },
            "SendEmailDto": {
                "type": "object",
                "properties": {}
            },
            "SendBulkEmailDto": {
                "type": "object",
                "properties": {}
            },
            "TrendDataPoint": {
                "type": "object",
                "properties": {
                    "date": {
                        "type": "string",
                        "description": "Date in YYYY-MM-DD format",
                        "example": "2025-11-03"
                    },
                    "users": {
                        "type": "number",
                        "description": "Total users on this date",
                        "example": 95
                    },
                    "sellers": {
                        "type": "number",
                        "description": "Total sellers on this date",
                        "example": 20
                    },
                    "products": {
                        "type": "number",
                        "description": "Total products on this date",
                        "example": 140
                    }
                },
                "required": [
                    "date",
                    "users",
                    "sellers",
                    "products"
                ]
            },
            "DashboardTrendsResponseDto": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "Success status",
                        "example": true
                    },
                    "data": {
                        "description": "Trend data points",
                        "example": [
                            {
                                "date": "2025-11-01",
                                "users": 95,
                                "sellers": 18,
                                "products": 140
                            },
                            {
                                "date": "2025-11-02",
                                "users": 97,
                                "sellers": 19,
                                "products": 145
                            },
                            {
                                "date": "2025-11-03",
                                "users": 100,
                                "sellers": 20,
                                "products": 150
                            }
                        ],
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TrendDataPoint"
                        }
                    },
                    "period": {
                        "type": "string",
                        "description": "Time period",
                        "example": "7days"
                    },
                    "startDate": {
                        "type": "string",
                        "description": "Start date",
                        "example": "2025-10-27"
                    },
                    "endDate": {
                        "type": "string",
                        "description": "End date",
                        "example": "2025-11-03"
                    }
                },
                "required": [
                    "success",
                    "data",
                    "period",
                    "startDate",
                    "endDate"
                ]
            },
...
            "CreateOrderDto": {
                "type": "object",
                "properties": {}
            },
...
        }
    }
}